core-service/requirements.txt
asgiref==3.11.0
djangorestframework==3.16.1
sqlparse==0.5.5
tzdata==2025.3
gunicorn==21.2.0
Django==4.2.7
gunicorn==21.2.0
psycopg2-binary==2.9.9
Django>=4.2
djangorestframework-simplejwt
cryptography
celery>=5.3
pika==1.3.2
django-cors-headers==4.3.1
requests==2.31.0
